+++ alicloud/extension_ots.go
@@ -1,0 +1,9 @@
+package alicloud
+
+type PrimaryKeyTypeString string
+
+const (
+	IntegerType = PrimaryKeyTypeString("Integer")
+	StringType  = PrimaryKeyTypeString("String")
+	BinaryType  = PrimaryKeyTypeString("Binary")
+)